refactor: move Pipe from view to controller
