+++ __tests__/DomainTest/controllers/modules/InputValidator/orderFormatTest.js
@@ -1,4 +1,4 @@
-import InputValidator from '../../../../../src/views/modules/InputValidator';
+import InputValidator from '../../../../../src/controllers/modules/InputValidator';
 
 describe('orderFormat()', () => {
   test.each([
